fix(review): wait for all reviewers before consensus when some are quota-exhausted
Classify quota/auth reviewer execution failures as UNAVAILABLE outcomes so multi-reviewer consensus waits for remaining reviewers instead of failing the parent session early.

Move review result tests into a sibling module to keep review_cmd_result.rs below the monolith token limit and bump the workspace version to satisfy release metadata gates.
